requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-bigquery
python-multipart
